+++ docs/cookbook/backend-api/adding-an-attribute-to-product-export.md
@@ -5,12 +5,12 @@ This short cookbook describes the steps that you need to do when you want to add
 Let's say you have added `extId` attribute to `Product` entity following [the cookbook](/docs/cookbook/adding-new-attribute-to-an-entity.md) and you want to include the attribute in the export as well.
 
 ## 1. Implement your own product transformer by extending `ApiProductTransformer`
-[`ApiProductTransformer`](/packages/backend-api/src/Controller/V1/ApiProductTransformer.php) is responsible for providing array of product data to backend API controllers.
+[`ApiProductTransformer`](/packages/backend-api/src/Controller/V1/Product/ApiProductTransformer.php) is responsible for providing array of product data to backend API controllers.
 You need to add your attribute to the array that is returned in `transform()` method.
 ```php
-namespace Shopsys\ShopBundle\Controller\Api\V1;
+namespace Shopsys\ShopBundle\Controller\Api\V1\Product;
 
-use Shopsys\BackendApiBundle\Controller\V1\ApiProductTransformer as BaseApiProductTransformer;
+use Shopsys\BackendApiBundle\Controller\V1\Product\ApiProductTransformer as BaseApiProductTransformer;
 use Shopsys\FrameworkBundle\Model\Product\Product;
 
 class ApiProductTransformer extends BaseApiProductTransformer
@@ -31,7 +31,7 @@ class ApiProductTransformer extends BaseApiProductTransformer
 
 ## 2. Add information about the class extension into the container configuration in [`services.yml`](/project-base/src/Shopsys/ShopBundle/Resources/config/services.yml)
 ```yaml
-Shopsys\BackendApiBundle\Controller\V1\ApiProductTransformer: '@Shopsys\ShopBundle\Controller\Api\V1\ApiProductTransformer'
+Shopsys\BackendApiBundle\Controller\V1\Product\ApiProductTransformer: '@Shopsys\ShopBundle\Controller\Api\V1\Product\ApiProductTransformer'
 ```
 
 ## Conclusion
